removed digitalWrite used for timing revised time related.  more unsigned long, less accumulated errors, faster execution of sequencing paths.  Check tx timeout in Tx state, reset in Rx state.  Single calculation of time elapsed for sequencer and tx timeout. Single variable for time of entry to state, including Tx state (for timeout calc). Simplify variable names related to timing. Make more variables global. (should be private to a class, but this is not coded c++ style.

+++ include/Config.h
@@ -12,7 +12,7 @@ struct sConfig_t {
   } Step[4];                // sequencer has 3 steps
   bool         RTSEnable;   // true enabled, false disabled
   bool         CTSEnable;   // true enabled, false disabled
-  unsigned int Timeout;     // sec, Tx timeout timer0 means disabled
+  uint16_t     Timeout;     // sec, Tx timeout time of 0 means disabled
   uint16_t     CRC16;       // check for valid configuration table
 }; 
 #endif
